azure_sounds/requirements.txt
PyQt5
python-dotenv
azure-cognitiveservices-speech
pyinstaller